Make sure AssetManager is initialized on startup
According to [1], declaring an AssetManager static will cause problems
on Android. Initializing the AssetManager as a member of a Kotlin
Singleton object is essentially the same. To ensure the assetManager
member is initialized and available when the application starts, it is
now declared with lateinit and initialized in the create() function of
the BeardBlasterGame class.

[1]: https://github.com/libgdx/libgdx/wiki/Managing-your-assets#creating-an-assetmanager

Closes #31